feat: Refactor Pet Sitter API endpoints and update front-end calls to improve organization and clarity, including new approval and rejection functionalities, and enhanced data fetching for sitter details and history
